# backend/openapi/swagger.yaml
basePath: /
definitions:
...
  consts.NodeAccessPerm:
    enum:
    - open
    - partial
    - closed
    type: string
    x-enum-comments:
      NodeAccessPermClosed: 完全禁止
      NodeAccessPermOpen: 完全开放
      NodeAccessPermPartial: 部分开放
    x-enum-descriptions:
    - 完全开放
    - 部分开放
    - 完全禁止
    x-enum-varnames:
    - NodeAccessPermOpen
    - NodeAccessPermPartial
    - NodeAccessPermClosed
  consts.NodePermName:
    enum:
    - visible
    - visitable
    - answerable
    type: string
    x-enum-comments:
      NodePermNameAnswerable: 可被问答
      NodePermNameVisible: 导航内可见
      NodePermNameVisitable: 可被访问
    x-enum-descriptions:
    - 导航内可见
    - 可被访问
    - 可被问答
    x-enum-varnames:
    - NodePermNameVisible
    - NodePermNameVisitable
    - NodePermNameAnswerable
  consts.NodeRagInfoStatus:
    enum:
    - PENDING
    - RUNNING
    - FAILED
    - SUCCEEDED
    - REINDEX
    type: string
    x-enum-comments:
      NodeRagStatusFailed: 处理失败
      NodeRagStatusPending: 等待处理
      NodeRagStatusReindexing: 重新索引中
      NodeRagStatusRunning: 正在进行处理（文本分割、向量化等）
      NodeRagStatusSucceeded: 处理成功
    x-enum-descriptions:
    - 等待处理
    - 正在进行处理（文本分割、向量化等）
    - 处理失败
    - 处理成功
    - 重新索引中
    x-enum-varnames:
    - NodeRagStatusPending
    - NodeRagStatusRunning
    - NodeRagStatusFailed
    - NodeRagStatusSucceeded
    - NodeRagStatusReindexing
...
  domain.BatchMoveReq:
    properties:
      ids:
        items:
          type: string
        type: array
      kb_id:
        type: string
      parent_id:
        type: string
    required:
    - ids
    - kb_id
    type: object
...
  domain.CreateNodeReq:
    properties:
      content:
        type: string
      content_type:
        type: string
      emoji:
        type: string
      kb_id:
        type: string
      name:
        type: string
      nav_id:
        type: string
      parent_id:
        type: string
      position:
        type: number
      summary:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.NodeType'
        enum:
        - 1
        - 2
    required:
    - kb_id
    - name
    - nav_id
    - type
    type: object
...
  domain.MoveNodeReq:
    properties:
      id:
        type: string
      kb_id:
        type: string
      next_id:
        type: string
      parent_id:
        type: string
      prev_id:
        type: string
    required:
    - id
    - kb_id
    type: object
...
  domain.NodeActionReq:
    properties:
      action:
        enum:
        - delete
        type: string
      ids:
        items:
          type: string
        type: array
      kb_id:
        type: string
    required:
    - action
    - ids
    - kb_id
    type: object
...
  domain.NodeGroupDetail:
    properties:
      auth_group_id:
        type: integer
      auth_ids:
        items:
          type: integer
        type: array
      kb_id:
        type: string
      name:
        type: string
      node_id:
        type: string
      perm:
        $ref: '#/definitions/consts.NodePermName'
    type: object
  domain.NodeListItemResp:
    properties:
      content_type:
        type: string
      created_at:
        type: string
      creator:
        type: string
      creator_id:
        type: string
      editor:
        type: string
      editor_id:
        type: string
      emoji:
        type: string
      id:
        type: string
      name:
        type: string
      nav_id:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      position:
        type: number
      publisher_id:
        type: string
      rag_info:
        $ref: '#/definitions/domain.RagInfo'
      status:
        $ref: '#/definitions/domain.NodeStatus'
      summary:
        type: string
      type:
        $ref: '#/definitions/domain.NodeType'
      updated_at:
        type: string
    type: object
  domain.NodeMeta:
    properties:
      content_type:
        type: string
      emoji:
        type: string
      summary:
        type: string
    type: object
  domain.NodePermissions:
    properties:
      answerable:
        allOf:
        - $ref: '#/definitions/consts.NodeAccessPerm'
        description: 可被问答
      visible:
        allOf:
        - $ref: '#/definitions/consts.NodeAccessPerm'
        description: 导航内可见
      visitable:
        allOf:
        - $ref: '#/definitions/consts.NodeAccessPerm'
        description: 可被访问
    type: object
  domain.NodeStatus:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-comments:
      NodeStatusDraft: 更新未发布
      NodeStatusPublished: 已发布
      NodeStatusUnreleased: 草稿
    x-enum-descriptions:
    - 草稿
    - 更新未发布
    - 已发布
    x-enum-varnames:
    - NodeStatusUnreleased
    - NodeStatusDraft
    - NodeStatusPublished
  domain.NodeSummaryReq:
    properties:
      ids:
        items:
          type: string
        type: array
      kb_id:
        type: string
    required:
    - ids
    - kb_id
    type: object
  domain.NodeType:
    enum:
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - NodeTypeFolder
    - NodeTypeDocument
...
  domain.PWResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.RagInfo:
    properties:
      message:
        type: string
      status:
        $ref: '#/definitions/consts.NodeRagInfoStatus'
      synced_at:
        type: string
    type: object
  domain.RecommendNodeListResp:
    properties:
      emoji:
        type: string
      id:
        type: string
      name:
        type: string
      nav_id:
        type: string
      nav_name:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      position:
        type: number
      recommend_nodes:
        items:
          $ref: '#/definitions/domain.RecommendNodeListResp'
        type: array
      summary:
        type: string
      type:
        $ref: '#/definitions/domain.NodeType'
    type: object
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.ShareNodeDetailItem:
    properties:
      children:
        items:
          $ref: '#/definitions/domain.ShareNodeDetailItem'
        type: array
      emoji:
        type: string
      id:
        type: string
      meta:
        $ref: '#/definitions/domain.NodeMeta'
      name:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      position:
        type: number
      type:
        $ref: '#/definitions/domain.NodeType'
      updated_at:
        type: string
    type: object
...
  domain.UpdateNodeReq:
    properties:
      content:
        type: string
      content_type:
        type: string
      emoji:
        type: string
      id:
        type: string
      kb_id:
        type: string
      name:
        type: string
      nav_id:
        type: string
      position:
        type: number
      summary:
        type: string
    required:
    - id
    - kb_id
    type: object
...
  github_com_chaitin_niuniu-wiki_api_node_v1.NodeListGroupNavResp:
    properties:
      count:
        type: integer
      is_released:
        type: boolean
      list:
        items:
          $ref: '#/definitions/domain.NodeListItemResp'
        type: array
      nav_id:
        type: string
      nav_name:
        type: string
      position:
        type: number
    type: object
...
  v1.NodeDetailResp:
    properties:
      content:
        type: string
      created_at:
        type: string
      creator_account:
        type: string
      creator_id:
        type: string
      editor_account:
        type: string
      editor_id:
        type: string
      id:
        type: string
      kb_id:
        type: string
      meta:
        $ref: '#/definitions/domain.NodeMeta'
      name:
        type: string
      nav_id:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      publisher_account:
        type: string
      publisher_id:
        type: string
      pv:
        type: integer
      status:
        $ref: '#/definitions/domain.NodeStatus'
      type:
        $ref: '#/definitions/domain.NodeType'
      updated_at:
        type: string
    type: object
  v1.NodeMoveNavReq:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
      kb_id:
        type: string
      nav_id:
        type: string
    required:
    - ids
    - kb_id
    - nav_id
    type: object
  v1.NodePermissionEditReq:
    properties:
      answerable_groups:
        description: 可被问答
        items:
          type: integer
        type: array
      ids:
        items:
          type: string
        type: array
      kb_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      visible_groups:
        description: 导航内可见
        items:
          type: integer
        type: array
      visitable_groups:
        description: 可被访问
        items:
          type: integer
        type: array
    required:
    - ids
    - kb_id
    type: object
  v1.NodePermissionEditResp:
    type: object
  v1.NodePermissionResp:
    properties:
      answerable_groups:
        description: 可被问答
        items:
          $ref: '#/definitions/domain.NodeGroupDetail'
        type: array
      id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      visible_groups:
        description: 导航内可见
        items:
          $ref: '#/definitions/domain.NodeGroupDetail'
        type: array
      visitable_groups:
        description: 可被访问
        items:
          $ref: '#/definitions/domain.NodeGroupDetail'
        type: array
    type: object
  v1.NodeRestudyReq:
    properties:
      kb_id:
        type: string
      node_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - kb_id
    - node_ids
    type: object
  v1.NodeRestudyResp:
    type: object
  v1.NodeStatsResp:
    properties:
      unpublished_count:
        description: 未发布的文档数
        type: integer
      unreleased_nav_count:
        description: 未发布目录数量
        type: integer
      unstudied_count:
        description: 未学习的文档数
        type: integer
    type: object
...
  v1.ShareNodeDetailResp:
    properties:
      content:
        type: string
      created_at:
        type: string
      creator_account:
        type: string
      creator_id:
        type: string
      editor_account:
        type: string
      editor_id:
        type: string
      id:
        type: string
      kb_id:
        type: string
      list:
        items:
          $ref: '#/definitions/domain.ShareNodeDetailItem'
        type: array
      meta:
        $ref: '#/definitions/domain.NodeMeta'
      name:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      publisher_account:
        type: string
      publisher_id:
        type: string
      pv:
        type: integer
      status:
        $ref: '#/definitions/domain.NodeStatus'
      type:
        $ref: '#/definitions/domain.NodeType'
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: niuniu-wiki API documentation
  title: niuniu-wiki API
  version: "1.0"
paths:
...
  /api/v1/node:
    post:
      consumes:
      - application/json
      description: Create Node
      parameters:
      - description: Node
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateNodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - bearerAuth: []
      summary: Create Node
      tags:
      - node
  /api/v1/node/action:
    post:
      consumes:
      - application/json
      description: Node Action
      parameters:
      - description: Action
        in: body
        name: action
        required: true
        schema:
          $ref: '#/definitions/domain.NodeActionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - bearerAuth: []
      summary: Node Action
      tags:
      - node
  /api/v1/node/batch_move:
    post:
      consumes:
      - application/json
      description: Batch Move Node
      parameters:
      - description: Batch Move Node
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.BatchMoveReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Batch Move Node
      tags:
      - node
  /api/v1/node/detail:
    get:
      consumes:
      - application/json
      description: Get Node Detail
      parameters:
      - in: query
        name: format
        type: string
      - in: query
        name: id
        required: true
        type: string
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.NodeDetailResp'
              type: object
      security:
      - bearerAuth: []
      summary: Get Node Detail
      tags:
      - node
    put:
      consumes:
      - application/json
      description: Update Node Detail
      parameters:
      - description: Node
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateNodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Update Node Detail
      tags:
      - node
  /api/v1/node/list:
    get:
      consumes:
      - application/json
      description: Get Node List
      parameters:
      - in: query
        name: kb_id
        required: true
        type: string
      - in: query
        name: nav_id
        type: string
      - in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.NodeListItemResp'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: Get Node List
      tags:
      - node
  /api/v1/node/list/group/nav:
    get:
      consumes:
      - application/json
      description: Get unpublished or unstudied document list grouped by nav
      parameters:
      - in: query
        name: kb_id
        required: true
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: nav_ids
        type: array
      - in: query
        name: search
        type: string
      - enum:
        - released
        - unpublished
        - unstudied
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_chaitin_niuniu-wiki_api_node_v1.NodeListGroupNavResp'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: Get Node List Grouped by Nav
      tags:
      - node
  /api/v1/node/move:
    post:
      consumes:
      - application/json
      description: Move Node
      parameters:
      - description: Move Node
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.MoveNodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Move Node
      tags:
      - node
  /api/v1/node/move/nav:
    post:
      consumes:
      - application/json
      description: Move node (and all its descendants if folder) to a different nav
      parameters:
      - description: Move Node Nav
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.NodeMoveNavReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Move Node to Nav
      tags:
      - node
  /api/v1/node/permission:
    get:
      consumes:
      - application/json
      description: 文档授权信息获取
      operationId: v1-NodePermission
      parameters:
      - in: query
        name: id
        required: true
        type: string
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.NodePermissionResp'
              type: object
      security:
      - bearerAuth: []
      summary: 文档授权信息获取
      tags:
      - NodePermission
  /api/v1/node/permission/edit:
    patch:
      consumes:
      - application/json
      description: 文档授权信息更新
      operationId: v1-NodePermissionEdit
      parameters:
      - description: para
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/v1.NodePermissionEditReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.NodePermissionEditResp'
              type: object
      security:
      - bearerAuth: []
      summary: 文档授权信息更新
      tags:
      - NodePermission
  /api/v1/node/recommend_nodes:
    get:
      consumes:
      - application/json
      description: Recommend Nodes
      parameters:
      - in: query
        name: kb_id
        required: true
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: nav_ids
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: node_ids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.RecommendNodeListResp'
                  type: array
              type: object
      security:
      - bearerAuth: []
      summary: Recommend Nodes
      tags:
      - node
  /api/v1/node/restudy:
    post:
      consumes:
      - application/json
      description: 文档重新学习
      operationId: v1-NodeRestudy
      parameters:
      - description: para
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/v1.NodeRestudyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.NodeRestudyResp'
              type: object
      security:
      - bearerAuth: []
      summary: 文档重新学习
      tags:
      - Node
  /api/v1/node/stats:
    get:
      consumes:
      - application/json
      description: Get Node Statistics
      parameters:
      - in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.NodeStatsResp'
              type: object
      security:
      - bearerAuth: []
      summary: Get Node Statistics
      tags:
      - node
  /api/v1/node/summary:
    post:
      consumes:
      - application/json
      description: Summary Node
      parameters:
      - description: Summary Node
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.NodeSummaryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Summary Node 异步后台生成
      tags:
      - node
  /api/v1/node/summary/stream:
    post:
      consumes:
      - application/json
      description: Stream Summary Node for single document
      parameters:
      - description: Summary Node
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.NodeSummaryReq'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
      security:
      - bearerAuth: []
      summary: Stream Summary Node
      tags:
      - node
...
  /share/v1/node/detail:
    get:
      consumes:
      - application/json
      description: GetNodeDetail
      parameters:
      - description: kb id
        in: header
        name: X-KB-ID
        required: true
        type: string
      - description: node id
        in: query
        name: id
        required: true
        type: string
      - description: format
        in: query
        name: format
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.ShareNodeDetailResp'
              type: object
      summary: GetNodeDetail
      tags:
      - share_node
  /share/v1/node/list:
    get:
      consumes:
      - application/json
      description: ShareNodeList
      parameters:
      - description: kb id
        in: header
        name: X-KB-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: ShareNodeList
      tags:
      - share_node
...
securityDefinitions:
  bearerAuth:
    description: Type "Bearer" + a space + your token to authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"